@@ -44,98 +43,90 @@
     "babel-cli": "^6.23.0",
     "babel-core": "^6.24.0",
     "babel-istanbul": "^0.12.2",
-    "babel-loader": "^6.4.1",
-    "babel-plugin-react-display-name": "^2.0.0",
-    "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
     "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-2": "^6.22.0",
-    "babel-watch": "^2.0.6",
     "chai": "^3.5.0",
+    "colors": "^1.1.2",
     "compression": "1.6.1",
-    "cross-env": "1.0.7",
     "css-loader": "0.27.3",
     "env2": "^2.1.1",
-    "enzyme": "^2.2.0",
-    "eslint": "^3.18.0",
-    "eslint-config-airbnb-base": "^11.1.0",
-    "eslint-import-resolver-webpack": "^0.8.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "5.0.1",
-    "eslint-watch": "^3.0.1",
-    "eventsource-polyfill": "0.9.6",
     "expect": "1.19.0",
-    "express": "4.13.4",
-    "extract-text-webpack-plugin": "1.0.1",
     "faker": "^4.1.0",
     "file-loader": "0.8.5",
-    "html-webpack-plugin": "^2.28.0",
+    "font-awesome": "^4.7.0",
     "istanbul": "^0.4.5",
     "jsdom": "8.5.0",
+    "material-icons": "^0.1.0",
     "mocha": "^2.4.5",
     "mocha-lcov-reporter": "^1.3.0",
     "nock": "8.0.0",
     "npm-run-all": "1.8.0",
     "open": "0.0.5",
-    "react-addons-test-utils": "^15.0.2",
-    "redux-immutable-state-invariant": "1.2.3",
-    "redux-mock-store": "1.0.2",
+    "prop-types": "^15.6.0",
     "rimraf": "2.5.2",
     "selenium-download": "^2.0.9",
     "selenium-standalone": "^6.2.0",
     "sinon": "^2.1.0",
     "style-loader": "0.13.1",
     "supertest": "^3.0.0",
-    "url-loader": "0.5.7",
-    "webpack-dev-middleware": "1.6.1",
-    "webpack-hot-middleware": "2.10.0"
+    "url-loader": "^0.6.2"
   },
   "dependencies": {
-    "axios": "0.13.1",
-    "babel-eslint": "^7.2.1",
-    "babel-polyfill": "6.8.0",
-    "babel-register": "^6.23.0",
+    "axios": "^0.17.1",
+    "babel-eslint": "^8.2.1",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-register": "^6.26.0",
     "bcrypt-nodejs": "0.0.3",
-    "body-parser": "^1.17.1",
-    "bootstrap": "3.3.6",
-    "cheerio": "^0.22.0",
+    "body-parser": "^1.18.2",
+    "cheerio": "^1.0.0-rc.2",
     "classnames": "^2.2.5",
-    "colors": "^1.1.2",
+    "clean-webpack-plugin": "^0.1.17",
     "dotenv": "^4.0.0",
-    "express": "^4.15.2",
-    "font-awesome": "^4.7.0",
-    "froala-editor": "^2.5.1",
-    "jquery": "2.2.3",
-    "jsonwebtoken": "^7.3.0",
+    "enzyme": "^3.3.0",
+    "eslint": "^4.15.0",
+    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint-import-resolver-webpack": "^0.8.4",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.5.1",
+    "express": "^4.16.2",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "froala-editor": "^2.7.3",
+    "html-webpack-plugin": "^2.30.1",
+    "jquery": "^3.3.0",
+    "jsonwebtoken": "^8.1.0",
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.4",
-    "material-icons": "^0.1.0",
     "materialize": "^1.0.0",
-    "materialize-css": "^0.98.1",
-    "morgan": "^1.8.1",
-    "nightwatch": "^0.9.14",
+    "materialize-css": "^0.100.2",
+    "morgan": "^1.9.0",
+    "nightwatch": "^0.9.19",
     "pg": "^6.1.4",
     "pg-hstore": "^2.3.2",
-    "react": "15.0.2",
-    "react-dom": "15.0.2",
-    "react-froala-wysiwyg": "^2.5.1",
-    "react-redux": "4.4.5",
-    "react-redux-sweetalert": "^1.0.0",
-    "react-router": "2.4.0",
-    "react-router-dom": "^4.0.0",
-    "react-router-redux": "4.0.4",
-    "react-rte": "^0.11.0",
-    "react-select": "^1.0.0-rc.3",
-    "redux": "^3.6.0",
-    "redux-devtools-extension": "^2.13.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-froala-wysiwyg": "^2.7.3",
+    "react-hot-loader": "^3.1.3",
+    "react-redux": "^5.0.6",
+    "react-redux-sweetalert": "^1.0.1",
+    "react-router-dom": "^4.2.2",
+    "redux": "^3.7.2",
+    "redux-devtools-extension": "^2.13.2",
+    "redux-immutable-state-invariant": "^2.1.0",
+    "redux-mock-store": "^1.4.0",
     "redux-router": "^2.1.2",
     "redux-thunk": "^2.2.0",
-    "sequelize": "^3.30.2",
-    "sequelize-cli": "^2.5.1",
+    "sequelize": "^4.31.2",
+    "sequelize-cli": "^3.2.0",
     "shortid": "^2.2.8",
     "toastr": "2.1.2",
-    "validator": "^7.0.0",
-    "webpack": "^1.14.0"
+    "validator": "^9.2.0",
+    "webpack": "^3.10.0",
+    "webpack-dev-middleware": "^2.0.4",
+    "webpack-dev-server": "^2.11.0",
+    "webpack-hot-middleware": "^2.21.0"
   }
 }
