@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import toastr from 'toastr';
 import { connect } from 'react-redux';
 import { deleteDocument } from '../../actions/documentActions';
@@ -67,9 +68,9 @@ class DocumentList extends Component {
 
 DocumentList.propTypes = {
   deleteDocument: PropTypes.func.isRequired,
-  docs: React.PropTypes.array.isRequired,
+  docs: PropTypes.array.isRequired,
   showModal: PropTypes.func.isRequired,
-  user: React.PropTypes.object.isRequired,
+  user: PropTypes.object.isRequired,
 };
 
 /**
