@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { addDisabledClass } from './disabledItems';
 
 
@@ -30,7 +31,7 @@ const RoleList = ({ roles, editRole, deleteRole }) =>
 RoleList.propTypes = {
   editRole: PropTypes.func.isRequired,
   deleteRole: PropTypes.func.isRequired,
-  roles: React.PropTypes.array.isRequired
+  roles: PropTypes.array.isRequired
 };
 
 
