@@ -1,31 +1,25 @@
-import React, { PropTypes } from 'react';
-import { Link, IndexLink } from 'react-router';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
-// import classname from 'classnames';
 import { logout } from '../../actions/authenticationAction';
 import { searchDocuments } from '../../actions/documentActions';
 
-export class Header extends React.Component {
-  constructor(props) {
-    super(props);
-    this.logout = this.logout.bind(this);
-    this.handleSearch = this.handleSearch.bind(this);
-  }
-
-  logout(e) {
+class Header extends Component {
+  logout = (e) => {
     e.preventDefault();
     this.props.logout();
   }
 
-  handleSearch(e) {
-    const path = this.props.location.pathname.slice(1);
+  handleSearch = (e) => {
+    const path = window.location.pathname.slice(1);
     if (['dashboard', 'documents'].includes(path)) {
       this.props.searchDocuments(e.target.value);
     }
   }
 
   getLinks({ isAuthenticated, user, isAdmin }) {
-    const path = this.props.location.pathname.slice(1);
+    const path = window.location.pathname.slice(1);
     const enabled = ['dashboard', 'documents'].includes(path);
     if (isAuthenticated) {
       return (
@@ -33,20 +27,35 @@ export class Header extends React.Component {
           <li>
             <form className="leftsearchbox">
               <div className="input-field">
-                <input disabled={!enabled} id="search" type="search" onChange={this.handleSearch} />
-                <label htmlFor="search"><i className="mdi mdi-magnify"></i></label>
+                <input
+                  disabled={!enabled}
+                  id="search"
+                  type="search"
+                  onChange={this.handleSearch}
+                />
+                <label
+                  htmlFor="search"
+                >
+                  <i className="mdi mdi-magnify" />
+                </label>
               </div>
             </form>
           </li>
-          <li><Link to="/dashboard" activeClassName="active">
-            <i className="material-icons left">dashboard</i>Dashboard</Link></li>
-          <li activeClassName="active">
+          <li>
+            <Link
+              to="/dashboard"
+              activeclassname="active"
+            >
+              <i className="material-icons left">dashboard</i>Dashboard
+            </Link>
+          </li>
+          <li activeclassname="active">
             <a href="#">Welcome, {user.userName}!</a>
           </li>
-          <li activeClassName="active">
+          <li activeclassname="active">
             <Link to="/profilepage">Profile</Link>
           </li>
-          <li activeClassName="active" id="personalDocs">
+          <li activeclassname="active" id="personalDocs">
             <Link to="/documents">Saved Documents</Link>
           </li>
           {isAdmin &&
@@ -58,15 +67,15 @@ export class Header extends React.Component {
             <Link to="/admin/handleusers">Manage Users</Link>
           </li>}
           <li>
-            <a href="#" activeClassName="active" onClick={this.logout}>Logout</a>
+            <a href="#" activeclassname="active" onClick={this.logout}>Logout</a>
           </li>
         </ul>
       );
     }
     return (
       <ul>
-        <li><Link to="/signup" activeClassName="active">Signup</Link></li>
-        <li><Link to="/login" activeClassName="active">Login</Link></li>
+        <li><Link to="/signup" activeclassname="active">Signup</Link></li>
+        <li><Link to="/login" activeclassname="active">Login</Link></li>
       </ul>
     );
   }
@@ -76,8 +85,9 @@ export class Header extends React.Component {
     return (
       <nav className="blue-grey">
         <div className="nav-wrapper">
-          <IndexLink to="/" activeClassName="active">
-            <i className="material-icons left">home</i>Home</IndexLink>
+          <Link to="/" activeclassname="active">
+            <i className="material-icons left">home</i>Home
+          </Link>
           <ul id="nav-mobile" className="right">
             <li>
               {navLinks}
@@ -91,17 +101,12 @@ export class Header extends React.Component {
 
 Header.propTypes = {
   user: PropTypes.object.isRequired,
-  location: PropTypes.object.isRequired,
   isAuthenticated: PropTypes.bool.isRequired,
   logout: PropTypes.func.isRequired,
   searchDocuments: PropTypes.func.isRequired,
   isAdmin: PropTypes.bool.isRequired
 };
 
-/**
- * @param {object}
- * @returns {object} data
- */
 export const mapStateToProps = (state) => {
   const { auth: { isAuthenticated, user } } = state;
   const isAdmin = isAuthenticated && user.userRoleId === 1;
@@ -112,4 +117,12 @@ export const mapStateToProps = (state) => {
   };
 };
 
-export default connect(mapStateToProps, { logout, searchDocuments })(Header);
+const enhance = connect(
+  mapStateToProps,
+  {
+    logout,
+    searchDocuments
+  }
+);
+
+export default enhance(Header);
