@@ -1,7 +1,7 @@
-import React from 'react';
+import React, { Component } from 'react';
 import LoginForm from './LoginForm';
 
-class LoginPage extends React.Component {
+class LoginPage extends Component {
   render() {
     return (
       <div className="card" id="logincard">
