@@ -1,10 +1,11 @@
-import React, { PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import SignupForm from './SignupForm';
 import { userSignupRequest, isUserExists } from '../../actions/signupActions';
 import { addFlashMessage } from '../../actions/flashMessages';
 
-class SignupPage extends React.Component {
+class SignupPage extends Component {
   render() {
     const { userSignupRequest, addFlashMessage, isUserExists } = this.props;
     return (
@@ -26,4 +27,10 @@ SignupPage.propTypes = {
   isUserExists: PropTypes.func.isRequired
 };
 
-export default connect(null, { userSignupRequest, addFlashMessage, isUserExists })(SignupPage);
+export default connect(
+  null,
+  {
+    userSignupRequest,
+    addFlashMessage,
+    isUserExists
+  })(SignupPage);
