@@ -1,18 +1,14 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import 'babel-polyfill';
 import React from 'react';
 import { render } from 'react-dom';
-import { Router, browserHistory } from 'react-router';
 import { Provider } from 'react-redux';
 import jwtDecode from 'jwt-decode';
 import { setCurrentUser } from './actions/authenticationAction';
 import configureStore from './store/configureStore';
-import routes from './routes';
+import App from './components/App';
 import setAuthorizationToken from './utils/setAuthorizationToken';
 import '../node_modules/materialize-css/dist/js/materialize.min';
 import '../node_modules/materialize-css/dist/css/materialize.min.css';
 import '../node_modules/material-icons/css/material-icons.css';
-import '../node_modules/sweetalert/dist/sweetalert.css';
 import './styles/styles.css';
 
 const store = configureStore();
@@ -24,6 +20,7 @@ if (localStorage.jwtToken) {
 
 render(
   <Provider store={store}>
-    <Router history={browserHistory} routes={routes} />
-  </Provider>, document.getElementById('app')
+    <App />
+  </Provider>,
+  document.getElementById('app')
 );
