@@ -1,12 +1,9 @@
 /* eslint-disable no-unused-expressions */
-
 import { expect } from 'chai';
 import model from '../../models';
 import helper from '../helper';
 
-const Role = model.Role;
-const User = model.User;
-const Document = model.Document;
+const { Role, User, Document } = model;
 
 const fakeUser = helper.createUser();
 const fakeDocument = helper.createDocument();
@@ -15,7 +12,7 @@ const requiredFields = ['title', 'content', 'access'];
 
 describe('The Document Model Test Suite', () => {
   describe('The Document Model', () => {
-    let document;
+    let documents;
     let owner;
 
     before((done) => {
@@ -23,6 +20,7 @@ describe('The Document Model Test Suite', () => {
         .then((createdRole) => {
           fakeUser.roleId = createdRole.id;
           return User.create(fakeUser);
+          done();
         })
         .then((createdUser) => {
           owner = createdUser;
@@ -31,8 +29,9 @@ describe('The Document Model Test Suite', () => {
         });
     });
 
-    beforeEach(() => {
-      document = Document.build(fakeDocument);
+    beforeEach((done) => {
+      documents = Document.build(fakeDocument);
+      done()
     });
 
     afterEach(() => Document.destroy({ where: {} }));
@@ -40,7 +39,7 @@ describe('The Document Model Test Suite', () => {
     after(() => model.sequelize.sync({ force: true }));
 
     it('should allow a document be created', (done) => {
-      document.save()
+      documents.save()
         .then((createdDocument) => {
           expect(createdDocument).to.exist;
           expect(typeof createdDocument).to.equal('object');
@@ -49,7 +48,7 @@ describe('The Document Model Test Suite', () => {
     });
 
     it('should create a document that has both title and content', (done) => {
-      document.save()
+      documents.save()
         .then((createdDocument) => {
           expect(createdDocument.title).to.equal(fakeDocument.title);
           expect(createdDocument.content).to.equal(fakeDocument.content);
@@ -58,7 +57,7 @@ describe('The Document Model Test Suite', () => {
     });
 
     it('should note the time the document was created', (done) => {
-      document.save()
+      documents.save()
         .then((createdDocument) => {
           expect(createdDocument.createdAt).to.exist;
           done();
@@ -67,7 +66,7 @@ describe('The Document Model Test Suite', () => {
 
     it('should have the access privilege of a created document specified',
       (done) => {
-        document.save()
+        documents.save()
           .then((createdDocument) => {
             expect(createdDocument.access).to.equal('public');
             done();
@@ -79,8 +78,8 @@ describe('The Document Model Test Suite', () => {
       creation`, () => {
         requiredFields.forEach((field) => {
           it(`requires a ${field} field to create a document`, () => {
-            document[field] = null;
-            return document.save()
+            documents[field] = null;
+            return documents.save()
               .catch((error) => {
                 expect(/notNull Violation/.test(error.message)).to.be.true;
               });
