@@ -1,16 +1,12 @@
 /* eslint-disable no-unused-vars */
-/* eslint-disable import/no-extraneous-dependencies */
-// More info on Webpack's Node API here: https://webpack.github.io/docs/node.js-api.html
-// Allowing console calls below since this is a build file.
 /* eslint-disable no-console */
 import webpack from 'webpack';
 import colors from 'colors';
+import webpackConfig from '../prod.config.babel';
 
-import webpackConfig from '../webpack.config.prod';
+process.env.NODE_ENV = 'production';
 
-process.env.NODE_ENV = 'production'; // this assures the Babel dev config (for hot reloading) doesn't apply.
-
-console.log('Generating minified bundle for production via Webpack. This will take a moment...'.blue);
+console.log('Generating minified bundle via Webpack...'.blue);
 
 webpack(webpackConfig).run((err, stats) => {
   if (err) { // so a fatal error occurred. Stop here.
@@ -32,9 +28,8 @@ webpack(webpackConfig).run((err, stats) => {
   console.log(`Webpack stats: ${stats}`);
 
   // if we got this far, the build succeeded.
-  console.log('Your app has been compiled in production mode and written to /dist. Ready to get cracking!'.green);
+  console.log('Compiled in production mode, written to /dist.'.green);
 
   return 0;
 });
-
 
