@@ -1,4 +1,6 @@
 /* eslint-disable import/no-extraneous-dependencies */
+/* eslint-disable import/no-dynamic-require */
+
 // This is the application entry point. server is set up here.
 import webpack from 'webpack';
 import path from 'path';
@@ -6,11 +8,12 @@ import express from 'express';
 import open from 'open';
 import compression from 'compression';
 import app from '../server/app';
-import config from '../webpack.config.dev';
+import config from '../dev.config.babel';
 
 const port = parseInt(process.env.PORT, 10) || 7070;
 
 /* eslint-disable no-console */
+/* eslint-disable global-require */
 const env = (process.env.NODE_ENV || 'development') === 'development';
 if (env) {
   const compiler = webpack(config);
@@ -26,7 +29,9 @@ if (env) {
 }
 
 app.get('*', (request, response) => {
-  response.sendFile(path.join(__dirname, `../${env ? 'client' : 'dist'}/index.html`));
+  response.sendFile(
+    path.join(__dirname, `../${env ? 'client' : 'dist'}/index.html`)
+  );
 });
 
 app.listen(port, (error) => {
